+++ docs/CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 For more details on new features, please check the [Manual](./MANUAL.md).
 
+<details><summary>7.5.5 - 13 February 2023</summary>
+
+### Added
+- The `[controlnet]` shortcode now supports the Normal Map model
+
+### Changed
+
+- Fixed another memory leak related to `[controlnet]`
+
+</details>
+
 <details><summary>7.5.4 - 13 February 2023</summary>
 
 ### Added